requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
PyMySQL
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-dotenv
python-multipart
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client